Modification de l'algorithme DQNagent.

+++ SnakeAI/src/headers/mlp.hpp
@@ -20,23 +20,23 @@ class MLP
      * @param inputs: std::vector<std::vector<double>>&, entrées pour lesquelles entrainer le réseau courant
      * @param targets: std::vector<std::vector<double>>&, sorties attendues du réseau
      * @param nb_iter: int, nombre d'itérations d'entrainement du réseau
-     * @param learning_rate: double, taux d'apprentissage du réseau
+     * @param learning_rate: float, taux d'apprentissage du réseau
     */
-    void train(std::vector<std::vector<double>>& inputs, std::vector<std::vector<double>>& targets, int nb_iter, double learning_rate);
+    void train(std::vector<std::vector<double>>& inputs, std::vector<std::vector<double>>& targets, int nb_iter, float learning_rate);
     
     /**
      * @brief Permet au réseau de prédire une valeur
      * @param inputs: vector<double>&, tableau contenant les entrées du réseau
-     * @return vector<double>&, tableau correspondant à la sortie du réseau
+     * @return vector<float>&, tableau correspondant à la sortie du réseau
     */
     std::vector<double>& feed_forward(std::vector<double>& inputs);
     
     /**
      * @brief Méthode permettant de faire apprendre le réseau tout entié
      * @param target: std::vector<double>, sortie voulu pour le réseau
-     * @param learning_rate: double, taux d'apprentissage du réseau
+     * @param learning_rate: float, taux d'apprentissage du réseau
      */
-    void backward_propagate(std::vector<double>& target, double learning_rate);
+    void backward_propagate(std::vector<double>& target, float learning_rate);
     
     /**
      * @brief Méthode renvoyant une chaine à afficher correspondant à la représentation du réseau
